pyproject.toml
[project]
name = "egretwind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.3",
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "mysqlclient>=2.2.7",
    "pyyaml>=6.0.3",
]
